My code suggestions
