+++ app/src/main/java/com/locoquest/app/MainActivity.kt
@@ -19,6 +19,13 @@ import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
+import com.locoquest.app.dao.IBenchmarkDAO
+import com.locoquest.app.dto.Benchmark
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
 
 
 private val TAG : String = MainActivity::class.java.name
@@ -138,4 +145,30 @@ class MainActivity : AppCompatActivity(), OnMapReadyCallback{
     override fun onMapReady(p0: GoogleMap) {
 
     }
+}
+fun main() {
+    val retrofit = Retrofit.Builder()
+        .baseUrl("https://geodesy.noaa.gov/api/nde/")
+        .addConverterFactory(MoshiConverterFactory.create())
+        .build()
+
+    val benchmarkDAO = retrofit.create(IBenchmarkDAO::class.java)
+
+    val pid = "AB1234"
+
+    val call = benchmarkDAO.getBenchmarkByPid(pid)
+    call.enqueue(object : Callback<Benchmark> {
+        override fun onResponse(call: Call<Benchmark>, response: Response<Benchmark>) {
+            if (response.isSuccessful) {
+                val benchmark = response.body()
+                println(benchmark)
+            } else {
+                println("Error: ${response.code()}")
+            }
+        }
+
+        override fun onFailure(call: Call<Benchmark>, t: Throwable) {
+            println("Error: ${t.message}")
+        }
+    })
 }
